// mocho-ui/src/components/form-fields/types.ts
export interface FormikFieldProps<T = Record<string, unknown>> {
  values: T;
  errors: FormikErrors<T>;
  touched: FormikTouched<T>;
  handleChange: (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => void;
  handleBlur: (e: React.FocusEvent<HTMLInputElement | HTMLTextAreaElement>) => void;
  setFieldValue: (field: string, value: unknown) => void;
}